requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
boto3==1.35.54
botocore==1.35.54
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
distro==1.9.0
dj-database-url==2.2.0
Django==4.2
django-cors-headers==3.13.0
django-storages==1.14.4
djangorestframework==3.15.2
drf-nested-routers==0.94.1
environ==1.0
exceptiongroup==1.2.2
frozenlist==1.4.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
jmespath==1.0.1
multidict==6.1.0
mysql-connector-python==9.0.0
mysqlclient==2.1.0
openai==0.28.0
packaging==24.1
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
railway==0.0.4
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.25.4
yarl==1.14.0
